1 file for campaings in params entities
